requirements.txt
flask
gunicorn
openai
pinecone
python-dotenv
requests
beautifulsoup4